feat: add solo round cutin component and related hooks
- Implemented `SoloRoundCutin` component for displaying round cut-ins with animations and themes.
- Created `WeatherDetailModal` for showing detailed weather information with animations.
- Added `useSoloCutins` hook for managing cut-in states and triggering animations.
- Developed `useSoloGame` hook for managing solo game state, including round submissions and game initialization.
